requirements.txt
fastapi
uvicorn
faster-whisper
python-dotenv
python-multipart